Fixed coordinate error
Fixed an error that caused the user input '[column, row]' make a component get placed at position '[row, column]' in the player's ship board. Most usages of 'GameUpdate' had no need for the field 'interestedPlayerId', so it was removed from the required fields and can now be set when necessary with 'setInterestedPlayerId(UUID)'.